NEW-209: create trainer button is display as part of the trainerlist header
