+++ amplify/backend/function/ec2start/ec2start-cloudformation-template.json
@@ -60,7 +60,7 @@
 				"Timeout": "25",
 				"Code": {
 					"S3Bucket": "newborn-app-20190515080104-deployment",
-					"S3Key": "amplify-builds/ec2start-1567181086-latest-build.zip"
+					"S3Key": "amplify-builds/ec2start-1567523484-latest-build.zip"
 				}
 			}
 		},
