+++ src/components/molecules/buttons/index.js
@@ -1,5 +1,6 @@
 import BadgeButton from "./badgeButton/badgeButton";
 import IconButton from "./iconButton/iconButton";
 import RoundButton from "./roundButton/roundButton";
+import Button from "./defaultButton/defaultButton";
 
-export { BadgeButton, IconButton, RoundButton };
+export { BadgeButton, IconButton, RoundButton, Button };
