+++ src/containers/trainer/list/trainerList.js
@@ -1,48 +1,21 @@
 import React from "react";
 import PropTypes from "prop-types";
-import gql from "graphql-tag";
-import { Mutation } from "react-apollo";
-import { withRouter } from "react-router-dom";
-
 import TrainerCard from "../../../components/organisms/cards/trainerCard/trainerCard";
 import CardList from "../../../components/organisms/lists/cardList/cardList";
-import { CreateTrainerCard } from "../../../components/organisms/cards/createTrainerCard";
 
-const CREATE_TRAINER = gql`
-  mutation CreateTrainer($title: String!, $trainerOwnerId: ID!) {
-    createTrainer(input: { title: $title, trainerOwnerId: $trainerOwnerId }) {
-      title
-    }
-  }
-`;
-
-function TrainerList({ history, trainers, title, loading, currentUserId }) {
+function TrainerList({ trainers, title, loading }) {
   let newbornCardList = [];
+
   if (loading) {
     for (let i = 0; i < 9; i++) {
       newbornCardList = [...newbornCardList, <TrainerCard loading={loading} />];
     }
   } else {
-    newbornCardList = trainers.map(trainer => {
+    const trainerList = trainers.map(trainer => {
       return <TrainerCard title={trainer.title} id={trainer.id} />;
     });
+    newbornCardList = [...newbornCardList, ...trainerList];
   }
-
-  newbornCardList.push(
-    <Mutation mutation={CREATE_TRAINER}>
-      {(createTrainer, { data }) => {
-        if (data && data.createTrainer) {
-          history.push("./builder");
-        }
-        return (
-          <CreateTrainerCard
-            currentUserId={currentUserId}
-            onCreateClick={createTrainer}
-          />
-        );
-      }}
-    </Mutation>
-  );
 
   return (
     <CardList
@@ -58,10 +31,9 @@ TrainerList.propTypes = {
   history: PropTypes.shape({
     push: PropTypes.func.isRequired
   }).isRequired,
-  currentUserId: PropTypes.string.isRequired,
   trainers: PropTypes.array.isRequired,
   title: PropTypes.string.isRequired,
   loading: PropTypes.bool.isRequired
 };
 
-export default withRouter(TrainerList);
+export default TrainerList;
