+++ src/containers/user/hub/userHub.js
@@ -1,18 +1,47 @@
 import React from "react";
 import PropTypes from "prop-types";
 import gql from "graphql-tag";
-import { Query } from "react-apollo";
+import { Query, Mutation } from "react-apollo";
+import { withRouter } from "react-router-dom";
 import * as queries from "../../../graphql/queries";
 import NewbornList from "../../newborn/list/newbornList";
 import TrainerList from "../../trainer/list/trainerList";
 import Text from "../../../components/atoms/text/text";
+
+import { Button } from "../../../components/molecules/buttons";
 import { ErrorDialog } from "../../../components/molecules/snackbars/errorSnackBar/style/error.style";
 
-function UserHub({ currentUserId }) {
+const CREATE_TRAINER = gql`
+  mutation CreateTrainer($title: String!, $trainerOwnerId: ID!) {
+    createTrainer(input: { title: $title, trainerOwnerId: $trainerOwnerId }) {
+      title
+    }
+  }
+`;
+
+function UserHub({ history, currentUserId }) {
   const returnListTitle = title => (
-    <Text size="large" weight="bold">
-      {title}
-    </Text>
+    <div style={{ display: "flex" }}>
+      <Text size="large" weight="bold">
+        {title}
+      </Text>
+      <Mutation mutation={CREATE_TRAINER}>
+        {(createTrainer, { data }) => {
+          if (data && data.createTrainer) {
+            history.push("./builder");
+          }
+          return <Button color="primary">Create new trainer</Button>;
+        }}
+      </Mutation>
+      <Mutation mutation={CREATE_TRAINER}>
+        {(createTrainer, { data }) => {
+          if (data && data.createTrainer) {
+            history.push("./builder");
+          }
+          return <Button color="secondary">Browse trainer</Button>;
+        }}
+      </Mutation>
+    </div>
   );
 
   const returnUserTrainers = () => {
@@ -88,4 +117,4 @@ UserHub.propTypes = {
   currentUserId: PropTypes.string.isRequired
 };
 
-export default UserHub;
+export default withRouter(UserHub);
